+++ src/main/java/com/cgr/base/domain/models/entity/EntitySourceFinance.java
@@ -12,6 +12,6 @@ public class EntitySourceFinance {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-     private Long source_financing_id;
+     private Integer source_financing_id;
      private String source_financing_name;
 }
